+++ icu.macro.d.ts
@@ -1,4 +1,4 @@
-import React from 'react';
+import { ReactElement } from 'react';
 import { Trans } from './';
 import type { Namespace, TypeOptions, i18n, ParseKeys } from 'i18next';
 
@@ -7,8 +7,8 @@ export { Trans };
 type _DefaultNamespace = TypeOptions['defaultNS'];
 declare module 'react-i18next/icu.macro' {
   export interface PluralSubProps<
-    Key extends ParseKeys<Ns, {}>,
-    Ns extends Namespace = _DefaultNamespace
+    Key extends ParseKeys<Ns, {}, ''>,
+    Ns extends Namespace = _DefaultNamespace,
   > {
     children?: never;
     i18nKey?: Key;
@@ -16,15 +16,19 @@ declare module 'react-i18next/icu.macro' {
     ns?: Ns;
     count: number;
     values?: {};
-    zero?: string | React.ReactElement;
-    one?: string | React.ReactElement;
-    two?: string | React.ReactElement;
-    few?: string | React.ReactElement;
-    many?: string | React.ReactElement;
-    other: string | React.ReactElement;
+    zero?: string | ReactElement;
+    one?: string | ReactElement;
+    two?: string | ReactElement;
+    few?: string | ReactElement;
+    many?: string | ReactElement;
+    other: string | ReactElement;
   }
 
-  type PluralProps<T, Key extends ParseKeys<Ns, {}>, Ns extends Namespace = _DefaultNamespace> = {
+  type PluralProps<
+    T,
+    Key extends ParseKeys<Ns, {}, ''>,
+    Ns extends Namespace = _DefaultNamespace,
+  > = {
     [P in keyof T]: P extends keyof PluralSubProps<Key, Ns>
       ? // support the standard properties of Plural
         PluralSubProps<Key, Ns>[P]
@@ -32,12 +36,12 @@ declare module 'react-i18next/icu.macro' {
       // technically it also supports $-1={..} and $2.3={..} but we don't need to
       // worry since that's invalid syntax.
       P extends `$${number}`
-      ? string | React.ReactElement
+      ? string | ReactElement
       : never;
   };
 
   interface SelectSubProps {
-    [key: string]: string | React.ReactElement;
+    [key: string]: string | ReactElement;
   }
 
   interface NoChildren {
@@ -45,41 +49,41 @@ declare module 'react-i18next/icu.macro' {
   }
 
   interface SelectRequiredProps<
-    Key extends ParseKeys<Ns, {}>,
-    Ns extends Namespace = _DefaultNamespace
+    Key extends ParseKeys<Ns, {}, ''>,
+    Ns extends Namespace = _DefaultNamespace,
   > extends NoChildren {
     i18nKey?: Key;
     i18n?: i18n;
     ns?: Ns;
-    other: string | React.ReactElement;
+    other: string | ReactElement;
   }
 
   // defining it this way ensures that `other` is always defined, but allows
   // unlimited other select types.
   type SelectProps<
-    Key extends ParseKeys<Ns, {}>,
-    Ns extends Namespace = _DefaultNamespace
+    Key extends ParseKeys<Ns, {}, ''>,
+    Ns extends Namespace = _DefaultNamespace,
   > = SelectSubProps & SelectRequiredProps<Key, Ns>;
 
-  function Plural<T, Key extends ParseKeys<Ns, {}>, Ns extends Namespace = _DefaultNamespace>(
+  function Plural<T, Key extends ParseKeys<Ns, {}, ''>, Ns extends Namespace = _DefaultNamespace>(
     props: PluralProps<T, Key, Ns> & NoChildren,
-  ): React.ReactElement;
+  ): ReactElement;
 
   function SelectOrdinal<
     T,
-    Key extends ParseKeys<Ns, {}>,
-    Ns extends Namespace = _DefaultNamespace
-  >(props: PluralProps<T, Key, Ns> & NoChildren): React.ReactElement;
+    Key extends ParseKeys<Ns, {}, ''>,
+    Ns extends Namespace = _DefaultNamespace,
+  >(props: PluralProps<T, Key, Ns> & NoChildren): ReactElement;
 
-  function Select<Key extends ParseKeys<Ns, {}>, Ns extends Namespace = _DefaultNamespace>(
+  function Select<Key extends ParseKeys<Ns, {}, ''>, Ns extends Namespace = _DefaultNamespace>(
     props: SelectProps<Key, Ns>,
-  ): React.ReactElement;
+  ): ReactElement;
 
   function date(strings: TemplateStringsArray, variable: Date): string;
   function time(strings: TemplateStringsArray, variable: Date): string;
   function number(strings: TemplateStringsArray, variable: number): string;
 
-  type ValidInterpolations = React.ReactElement | string;
+  type ValidInterpolations = ReactElement | string;
 
   function plural(
     strings: TemplateStringsArray,