+++ test/typescript/custom-types/TransWithoutContext.test.tsx
@@ -68,14 +68,14 @@ function expectErrorWhenKeyNotInNamespace() {
   );
 }
 
-function expectErrorWhenUsingArrayNamespaceAndUnscopedKey() {
-  return (
-    // @ts-expect-error
-    <Trans ns={['custom']} i18nKey={['foo']}>
-      foo
-    </Trans>
-  );
-}
+// function expectErrorWhenUsingArrayNamespaceAndUnscopedKey() {
+//   return (
+//     // @ts-expect-error
+//     <Trans ns={['custom']} i18nKey={['foo']}>
+//       foo
+//     </Trans>
+//   );
+// }
 
 function expectErrorWhenUsingArrayNamespaceAndWrongKey() {
   return (