+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 13.0.0
+
+- Update types to support t function redesign [1615](https://github.com/i18next/react-i18next/pull/1615)
+- requires i18next >= v23.0.1
+
 ### 12.3.1
 
 - optimization for optional lng prop for useTranslation, should now prevent missings when lazy loading translations [1637](https://github.com/i18next/react-i18next/issues/1637)